Replace body with div container for buttons
The body was initially being used for the containment of the buttons
(and generally everything on the page). For better control of both the
button's contents as well as to add other elements on the page (score
display) we need to be able to add nodes which are non-siblings to the
buttons. This is accomplished as the buttons are now being contained in
their own div "container" and a sibling to container can be added for
score functionality without inheretting rules for the buttons.